requirements.txt
cycler==0.10.0
gitdb==4.0.5
GitPython==3.1.3
kiwisolver==1.2.0
matplotlib==3.2.2
numpy==1.19.0
pandas==1.0.5
pkg-resources==0.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
smmap==3.0.4
